Restructure to two separate external calls
Introduce a new call, `TrackNonCustomerEventAsync` to track invitation
events.

The reason is the result of calling the `TrackEventAsync` method could
be confusing if you have a customer factory hooked in. With a customer
factory, passing null for customer ID looks up the customer ID using the
factory, and will not send an invitation event.

This means we now have two methods:

- `TrackEventAsync` for regular events
- `TrackNonCustomerEventAsync` for invitation events